Refactor gulp tasks into separate files
