+++ gulp/tasks/build-safari.js
@@ -1,0 +1,6 @@
+// Here would be a good place to build the Safari archive. But it requires a
+// custom build of the `xar` executable to extract certificates from a
+// Safari-built .safariextz, then signing the .xar archive (renamed .safariextz)
+
+// Steps to build from the command line:
+// http://developer.streak.com/2013/01/how-to-build-safari-extension-using.html
